refactor(#101) :: common 도메인 구조 정리
