+++ src/test/java/com/example/gujeuck_server/domain/common/service/funnel/CreateCheckInFunnelEventServiceTest.java
@@ -1,9 +1,9 @@
-package com.example.gujeuck_server.domain.common.analytics.service;
+package com.example.gujeuck_server.domain.common.service.funnel;
 
-import com.example.gujeuck_server.domain.common.analytics.domain.enums.CheckInFunnelEventName;
-import com.example.gujeuck_server.domain.common.analytics.domain.repository.CheckInFunnelEventRepository;
-import com.example.gujeuck_server.domain.common.analytics.presentation.dto.request.CheckInFunnelEventRequest;
-import com.example.gujeuck_server.domain.common.analytics.presentation.dto.request.CheckInFunnelEventsRequest;
+import com.example.gujeuck_server.domain.common.domain.funnel.enums.CheckInFunnelEventName;
+import com.example.gujeuck_server.domain.common.domain.funnel.repository.CheckInFunnelEventRepository;
+import com.example.gujeuck_server.domain.common.presentation.funnel.dto.request.CheckInFunnelEventRequest;
+import com.example.gujeuck_server.domain.common.presentation.funnel.dto.request.CheckInFunnelEventsRequest;
 import com.example.gujeuck_server.domain.log.domain.repository.LogRepository;
 import com.example.gujeuck_server.domain.user.domain.enums.Age;
 import org.junit.jupiter.api.Test;
